+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [#107]: added Czech translation
 - [#112]: improved Dutch translation
 - [#117]: method `getFormOptions` returns an array for the `validation_groups` option
+- [#125]: added generic form options to simplify passing options to all steps
 
 [#98]: https://github.com/craue/CraueFormFlowBundle/issues/98
 [#101]: https://github.com/craue/CraueFormFlowBundle/issues/101
@@ -18,6 +19,7 @@
 [#107]: https://github.com/craue/CraueFormFlowBundle/issues/107
 [#112]: https://github.com/craue/CraueFormFlowBundle/issues/112
 [#117]: https://github.com/craue/CraueFormFlowBundle/issues/117
+[#125]: https://github.com/craue/CraueFormFlowBundle/issues/125
 
 ## 2.1.4 (2013-12-05)